done trang thai bill

+++ src/main/java/com/example/tmdt/service/impl/BillDetailService.java
@@ -52,13 +52,6 @@ public class BillDetailService implements IBillDetailService {
 
     @Override
     public void delete(Long id) {
-        Optional<BillDetail> billDetailOptional = billDetailRepository.findById(id);
-        if (billDetailOptional.isPresent()) {
-            BillDetail billDetail = billDetailOptional.get();
-            billDetail.getBill().setStatus("Đơn hủy");
-            billDetailRepository.save(billDetail);
-        }
-
 
     }
 
@@ -149,36 +142,38 @@ public class BillDetailService implements IBillDetailService {
         User user = userRepository.findUserByAccount_Id(idAccount);
         Optional<Bill> billOptional = billRepository.findBillByIdAccount(cartDetail.getCart().getAccount().getId(), cartDetail.getProduct().getShop().getId());
         Bill bill;
-        if (billOptional.isPresent() && billOptional.get().getStatus().equals("Chờ xác nhận")) {
-            bill = billOptional.get();
-        } else {
-            bill = new Bill();
-            bill.setAccount(cartDetail.getCart().getAccount());
-            bill.setShop(cartDetail.getProduct().getShop());
-            bill.setName(user.getName());
-            bill.setPhone(user.getPhone());
-            bill.setAddress(user.getAddress());
-            bill.setWards(user.getWards());
-            bill.setDate(LocalDate.now());
-            bill.setStatus("Chờ xác nhận");
-            billRepository.save(bill);
-        }
-        BillDetail billDetail = new BillDetail();
-        billDetail.setBill(bill);
-        billDetail.setProduct(cartDetail.getProduct());
-        billDetail.setQuantity(cartDetail.getQuantity());
-        Double newPrice = cartDetail.getProduct().getPrice() - (cartDetail.getProduct().getPrice() * cartDetail.getProduct().getPromotion() / 100);
-        billDetail.setPrice(newPrice);
-        Double total = cartDetail.getQuantity() * newPrice;
-        billDetail.setTotal(total);
-        Double quantity = cartDetail.getQuantity();
-        Product product = cartDetail.getProduct();
-        if (quantity <= product.getQuantity()) {
-            productRepository.save(product);
-            billDetailRepository.save(billDetail);
-        }
-        cartDetailRepository.deleteCartDetailByProduct(product.getId());
+        if (user.getAddress() != null && user.getPhone() != null) {
+            if (billOptional.isPresent() && billOptional.get().getStatus().equals("Chờ xác nhận")) {
+                bill = billOptional.get();
+            } else {
+                bill = new Bill();
+                bill.setAccount(cartDetail.getCart().getAccount());
+                bill.setShop(cartDetail.getProduct().getShop());
+                bill.setName(user.getName());
+                bill.setPhone(user.getPhone());
+                bill.setAddress(user.getAddress());
+                bill.setWards(user.getWards());
+                bill.setDate(LocalDate.now());
+                bill.setStatus("Chờ xác nhận");
+                billRepository.save(bill);
+            }
+            BillDetail billDetail = new BillDetail();
+            billDetail.setBill(bill);
+            billDetail.setProduct(cartDetail.getProduct());
+            billDetail.setQuantity(cartDetail.getQuantity());
+            Double newPrice = cartDetail.getProduct().getPrice() - (cartDetail.getProduct().getPrice() * cartDetail.getProduct().getPromotion() / 100);
+            billDetail.setPrice(newPrice);
+            Double total = cartDetail.getQuantity() * newPrice;
+            billDetail.setTotal(total);
+            Double quantity = cartDetail.getQuantity();
+            Product product = cartDetail.getProduct();
+            if (quantity <= product.getQuantity()) {
+                productRepository.save(product);
+                billDetailRepository.save(billDetail);
+            }
+            cartDetailRepository.deleteCartDetailByProduct(product.getId());
 
+        }
     }
